* commit '5d365a8898e08bc961caa93675d38c04ffa8d38b':
  GPLIM-6145 add space
  GPLIM-6145 include date way in the future matching error found in GPLIM-6143
  GPLIM-6145 DateUtils.getNumDaysBetween(d1,d2) ignores Month, Year, etc, units
  GPLIM-6145 DateUtils.getNumDaysBetween(d1,d2) ignores Month, Year, etc, units

+++ mercury/src/main/java/org/broadinstitute/gpinformatics/athena/presentation/orders/ProductOrderActionBean.java
@@ -771,7 +771,7 @@ public class ProductOrderActionBean extends CoreActionBean {
                 quote.get().getFunding().stream()
                     .filter(Funding::isFundsReservation)
                     .forEach(funding -> {
-                        int numDaysBetween = DateUtils.getNumDaysBetween(new Date(), funding.getGrantEndDate());
+                        long numDaysBetween = DateUtils.getNumDaysBetween(new Date(), funding.getGrantEndDate());
                         if (numDaysBetween > 0 && numDaysBetween < 45) {
                             addMessage(
                                 String.format("The Funding Source %s on %s  Quote expires in %d days. If it is likely "
